requirements.txt
aiofiles==0.6.0
aiosqlite==0.17.0
asyncua==0.9.12
bernhard==0.2.6
bottle==0.12.19
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
chevron==0.14.0
cothread==2.17
cryptography==3.4.7
docker==5.0.0
epicsdbbuilder==1.4
future==0.18.2
Glances==3.1.7
idna==2.10
ifaddr==0.1.7
lxml==4.6.3
netifaces==0.11.0
numpy==1.20.3
opcua==0.98.13
pandas==1.2.5
pbkdf2==1.3
ply==3.11
protobuf==3.17.3
psutil==5.8.0
py-cpuinfo==4.0.0
py3nvml==0.2.6
pyasn1==0.4.8
pycparser==2.20
pycryptodomex==3.10.1
pyepics==3.5.0
pygal==2.4.0
pymdstat==0.4.2
pySMART.smartx==0.3.10
pysmi==0.3.4
pysnmp==4.4.12
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
scandir==1.10.0
six==1.16.0
sortedcontainers==2.3.0
sparklines==0.4.2
typing-extensions==3.10.0.0
urllib3==1.26.5
websocket-client==1.1.0
wifi==0.3.8
xmltodict==0.12.0
zeroconf==0.31.0
